mobile view supported, cancellation exceptions caught

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, request, redirect, jsonify, send_file, send_from_directory, Response
+from flask import Flask, request, jsonify, send_from_directory, Response, session
 from flask_cors import CORS
 from werkzeug.utils import secure_filename
 from uuid import uuid4
@@ -7,16 +7,20 @@ import time
 from convert import ConvertImageProcess, ConvertVideoProcess
 
 app = Flask(__name__)
+app.secret_key = "secret"
 app.config.from_object("config.Config")
 config = app.config
 cors = CORS(app)
 
 jobs = {}
 max_jobs = config["MAX_JOBS"]
-progress_update_rate = 0.5
+jobs_count = 0
+progress_update_rate = config["PROGRESS_RATE"]
 
 @app.route("/api/convert", methods=["POST"])
 def convert():
+    print ("-" * 20)
+    print ("- Job request received")
     data = request.form
     fileUpload = request.files["fileUpload"]
     filename, file_ext = os.path.splitext(fileUpload.filename)
@@ -28,10 +32,12 @@ def convert():
     if len(jobs) >= max_jobs:
         print ("Too many jobs")
         return jsonify("max"), 200
-
+    
+    global jobs_count
+    jobs_count += 1
     if file_ext in config["IMG_EXT"]:
         p = ConvertImageProcess(
-            temp_path, output_path, override=False,
+            temp_path, output_path + ".jpg", override=False,
             image_reducer = int(data["imageReduction"]),
             fontSize = int(data["fontSize"]),
             spacing = float(data["spacing"]),
@@ -45,7 +51,7 @@ def convert():
     elif file_ext in config["VID_EXT"]:
         temp_batch_folder = os.path.join(config["TEMP"], file_id + "/")
         p = ConvertVideoProcess(
-            temp_path, output_path,
+            temp_path, output_path + ".mp4",
             temp_folder=temp_batch_folder,
             frame_frequency=int(data["frameFrequency"]),
             image_reducer = int(data["imageReduction"]),
@@ -63,10 +69,13 @@ def convert():
         return jsonify("Bad format")
 
 def cancel(job_id):
-    print ("Cancelled")
-    jobs[job_id].terminate_process()
+    print ("- Cancelling")
     terminated = jobs.pop(job_id, None)
-    job_type_ext = os.path.splitext(terminated.output)[1]
+    if terminated is None:
+        print (job_id, "is not running anymore!")
+        return False
+    terminated.terminate_process()
+    job_type_ext = os.path.splitext(terminated.output_path)[1]
     if job_type_ext == ".mp4":
         os.remove(terminated.video_path)
         terminated.cleanup_temp()
@@ -74,11 +83,12 @@ def cancel(job_id):
         os.remove(terminated.image_path)
     if os.path.isfile(terminated.output_path):
         os.remove(terminated.output_path)
+    print ("- Cancelled")
     return True
 
 @app.route("/api/getprogress", methods=["POST"])
 def get_progress():
-    print ("Getting progress")
+    print ("- Getting progress")
     job_id = request.get_json()
     job = jobs[job_id]
     def progress_stream():
@@ -88,18 +98,18 @@ def get_progress():
                 time.sleep(progress_update_rate)
         except GeneratorExit:
             if job.get_progress() < 100 and job_id in jobs:
-                print ("Client disconnected from progress stream")
+                print ("- Client disconnected from progress stream")
                 cancel(job_id)
     return Response(progress_stream(), mimetype="text/event-stream")
 
 @app.route("/api/getoutput", methods=["POST"])
 def get_output():
-    print ("Getting output")
+    print ("- Getting output")
     job_id = request.get_json()
     done = jobs.pop(job_id, None)
-    job_type_ext = os.path.splitext(done.output)[1]
+    job_type_ext = os.path.splitext(done.output_path)[1]
     os.remove(done.video_path if job_type_ext == ".mp4" else done.image_path)
-    return send_from_directory(config["OUTPUT"], job_id + ".txt" + job_type_ext, as_attachment=True), 200
+    return send_from_directory(config["OUTPUT"], job_id + job_type_ext, as_attachment=True), 200
 
 @app.route("/api/cancel", methods=["POST"])
 def cancel_conversion():
@@ -109,5 +119,7 @@ def cancel_conversion():
 
 if __name__ == "__main__":
     print ("=" * 50)
+    if not os.path.isdir("./temp"):
+        os.mkdir("./temp")
     app.run(host="0.0.0.0", port=5000, debug=1)
